refactor: rename feature for consistency
