+++ packages/mcp-server/src/features/source-documents/index.ts
