+++ packages/mcp-server/src/features/source-documents/utils/sanitize.test.ts
