+++ packages/shared/src/types/index.ts
@@ -3,7 +3,7 @@ export * as SmartConnections from "./plugin-smart-connections";
 export * as Templater from "./plugin-templater";
 export * from "./prompts";
 export * from "./smart-search";
-export * from "./source-document";
+export * from "./source-documents";
 
 export type SetupFunctionResult = Promise<
   { success: true } | { success: false; error: string }
